Move Frequency subclasses to separate files
