+++ lib/montrose/frequency/daily.rb
@@ -1,0 +1,9 @@
+module Montrose
+  class Frequency
+    class Daily < Frequency
+      def include?(time)
+        matches_interval? time.to_date - @starts.to_date
+      end
+    end
+  end
+end
