+++ lib/montrose/frequency/minutely.rb
@@ -1,0 +1,9 @@
+module Montrose
+  class Frequency
+    class Minutely < Frequency
+      def include?(time)
+        matches_interval?((time - @starts) / 1.minute)
+      end
+    end
+  end
+end
