+++ lib/montrose/frequency/monthly.rb
@@ -1,0 +1,9 @@
+module Montrose
+  class Frequency
+    class Monthly < Frequency
+      def include?(time)
+        matches_interval?((time.month - @starts.month) + (time.year - @starts.year) * 12)
+      end
+    end
+  end
+end
