+++ lib/montrose/frequency/yearly.rb
@@ -1,0 +1,9 @@
+module Montrose
+  class Frequency
+    class Yearly < Frequency
+      def include?(time)
+        matches_interval? time.year - @starts.year
+      end
+    end
+  end
+end
